// client/src/components/accordion/external/external-accordion.model.ts
export enum ExternalAccordionStructureEnum {
    Accordion = 'external-accordion',
    Section = 'external-accordion-section',
    Header = 'external-accordion-header',
    Content = 'external-accordion-content',
}

export enum ExternalAccordionDataAttrEnum {
    DataSectionActive = 'data-section-active',
}